api/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
djangorestframework==3.11.0
idna==2.10
multidict==4.7.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vcrpy==4.0.2
vcrpy-unittest==0.1.7
wrapt==1.12.1
yarl==1.4.2
